renamed supabase auth client file name
